Adding accounts, moving bin2 over top of bin, and adding in basic web3 provider
