+++ commands/accounts.js
@@ -1,8 +1,12 @@
-module.exports = async (params)=>{
-  const DEBUG = params.config.DEBUG;
-  if(DEBUG) console.log(" >>> ACCOUNTS")
-  if(DEBUG) console.log("Reading Accounts...")
+const fs = require('fs');
+const winston = require('winston');
+
+module.exports = async (params) => {
+  winston.debug(' >>> ACCOUNTS')
+  winston.debug('Reading Accounts...')
+
   let accounts = await params.web3.eth.getAccounts()
-  params.fs.writeFileSync("accounts.json",JSON.stringify(accounts))
+  fs.writeFileSync('accounts.json', JSON.stringify(accounts))
+
   return accounts
 }
